feat: moved PHD client to services folder and added logging on fetch failure
